refactor: move some RE functions to separate files
