+++ app/components/RichEditor/slateNodes.js
@@ -1,0 +1,30 @@
+export const DEFAULT_BLOCK = 'paragraph'
+export const BLOCK_BLOCKQUOTE = 'block-quote'
+export const BLOCK_BULLETED_LIST = 'bulleted-list'
+export const BLOCK_CODE = 'code'
+export const BLOCK_HEADING1 = 'heading1'
+export const BLOCK_HEADING2 = 'heading2'
+export const BLOCK_HEADING3 = 'heading3'
+export const BLOCK_HEADING4 = 'heading4'
+export const BLOCK_HEADING5 = 'heading5'
+export const BLOCK_HEADING6 = 'heading6'
+export const BLOCK_HR = 'horizontal-rule'
+export const BLOCK_IMAGE = 'image'
+export const BLOCK_LIST_ITEM = 'list-item'
+export const BLOCK_LIST_ITEM_TEXT = 'list-item-text'
+export const BLOCK_NUMBERED_LIST = 'ordered-list'
+
+export const HEADINGS = [
+  BLOCK_HEADING1,
+  BLOCK_HEADING2,
+  BLOCK_HEADING3,
+  BLOCK_HEADING4,
+  BLOCK_HEADING5,
+  BLOCK_HEADING6
+]
+
+export const INLINE_LINK = 'link'
+
+export const MARK_BOLD = 'bold'
+export const MARK_CODE = 'code'
+export const MARK_ITALIC = 'italic'
